refactor: extract ExifFormatters and FlowLayout from ExifPanelView
ExifPanelView.swift mixed three responsibilities — the SwiftUI view, a
half-dozen string formatters, and a generic FlowLayout. Splitting them
into focused files makes the formatters unit-testable in isolation and
lets FlowLayout be reused without dragging the view along.

- ExifFormatters.swift: pure Foundation enum namespace for
  formatNumber/formatExposure/formatDate/formatLocation/formatCoordinates.
- FlowLayout.swift: the keyword-chip layout, verbatim move.
- ExifPanelView.swift: now just the view (sectionHeader, exifRow, body).
- ExifFormattersTests.swift: covers all five formatters.
